got rid of errors
