+++ app/main.py
@@ -17,7 +17,7 @@ from app.api import token_endpoint, rule_endpoint, resource_endpoint, event_endp
     settings_endpoint, event_source_endpoint, test_endpoint, \
     event_tag_endpoint, consent_type_endpoint, flow_action_endpoint, flows_endpoint, info_endpoint, \
     user_endpoint, event_schema_validation_endpoint, debug_endpoint, log_endpoint, tracardi_pro_endpoint, \
-    storage_endpoint, destination_endpoint, user_log_endpoint
+    storage_endpoint, destination_endpoint, user_log_endpoint, user_account_endpoint
 from app.api.graphql.profile import graphql_profiles
 from app.api.scheduler import tasks_endpoint
 from app.api.track import event_server_endpoint
@@ -164,6 +164,7 @@ application.include_router(tracardi_pro_endpoint.router)
 application.include_router(storage_endpoint.router)
 application.include_router(destination_endpoint.router)
 application.include_router(user_log_endpoint.router)
+application.include_router(user_account_endpoint.router)
 
 # GraphQL
